requirements.txt
attr==0.3.2
bpython==0.23
brotli==1.0.9
brotlicffi==1.0.9.2
click==8.1.3
ConfigParser==5.3.0
cryptography==37.0.4
cx_Oracle==8.3.0
Cython==0.29.32
djangorestframework==3.14.0
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
ipdb==0.13.9
ipython==8.5.0
ipywidgets==8.0.2
jnius==1.1.0
keyring==23.9.3
lockfile==0.12.2
mock==4.0.3
MySQL-python==1.2.5
numpy==1.23.3
Pillow==9.2.0
protobuf==4.21.6
psycopg2==2.9.3
pylibmc==1.6.3
pymemcache==3.5.2
pyOpenSSL==22.0.0
pytz==2022.2.1
pywatchman==1.4.1
PyYAML==6.0
railroad==0.5.0
simplejson==3.17.6
Sphinx==5.2.0
tblib==1.7.0
toml==0.10.2
tornado==6.2
trove_classifiers==2022.8.31
truststore==0.5.0
unicodedata2==14.0.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
django-shortcuts
python-dotenv
